fix code following pr #32

+++ src/setup/tasks/delete-task-handler.js
@@ -2,46 +2,53 @@ import inquirer from 'inquirer';
 import colors from 'colors/safe';
 import tasksManager from '../../helpers/tasks-manager';
 
-class DeleteTaskHandler {
-  static async createAdapter() {
-    const answers = await inquirer.prompt([
-      {
-        type: 'list',
-        name: 'taskName',
-        message: 'Select a task to delete',
-        choices: [...await tasksManager.getTasksList(), {
-          name: 'Go Back',
-          value: '',
-        }],
-      },
-      {
-        type: 'confirm',
-        name: 'confirmDelete',
-        message: 'Are you sure?',
-        default: false,
-      },
-    ]);
+const DeleteTaskHandler = (function createDeleteTaskHandler() {
+  const _private = new WeakMap();
 
-    if (answers.confirmDelete) {
-      return new DeleteTaskHandler(answers.taskName);
-    }
-    console.log(colors.notify('Delete task cancelled'));
+  class DeleteTaskHandler {
+    static async createAdapter() {
+      const answers = await inquirer.prompt([
+        {
+          type: 'list',
+          name: 'taskName',
+          message: 'Select a task to delete',
+          choices: [...await tasksManager.getTasksList(), {
+            name: 'Go Back',
+            value: '',
+          }],
+        },
+        {
+          type: 'confirm',
+          name: 'confirmDelete',
+          message: 'Are you sure?',
+          when: answers => !!answers.taskName,
+          default: false,
+        },
+      ]);
 
-    return null;
-  }
+      if (answers.taskName && answers.confirmDelete) {
+        return new DeleteTaskHandler(answers.taskName);
+      }
+      console.log(colors.notify('Delete task cancelled'));
 
-  constructor(taskName) {
-    this._taskName = taskName;
-  }
-
-  async run() {
-    if (!this._taskName) {
-      throw new Error('Missing task name');
+      return null;
     }
 
-    console.log(colors.notify(`Task '${this._taskName}' deleted`));
-    await tasksManager.deleteTask(this._taskName);
+    constructor(taskName) {
+      _private.set(this, { taskName });
+    }
+
+    async run() {
+      if (!_private.get(this).taskName) {
+        throw new Error('Missing task name');
+      }
+
+      console.log(colors.notify(`Task '${_private.get(this).taskName}' deleted`));
+      await tasksManager.deleteTask(_private.get(this).taskName);
+    }
   }
-}
+
+  return DeleteTaskHandler;
+}());
 
 export default DeleteTaskHandler;
